+++ main/peripherals/NTC100k.h
@@ -59,7 +59,7 @@ class NTC100k {
       }
     }
 
-    const int getTemperature() {
+    const float getTemperature() {
       return temperature;
     }
 };
